mCommandSelected can be set to -1, so changed to be an int.  Also some casting to avoid warnings about int/unsigned mismatches.

+++ src/prefs/KeyConfigPrefs.h
@@ -55,7 +55,7 @@ class KeyConfigPrefs:public PrefsPanel
    wxListCtrl *mList;
 
    CommandManager *mManager;
-   size_t mCommandSelected;
+   int mCommandSelected;
 
    wxArrayString mCats;
    wxArrayString mNames;
